BUG GAME-003 — preserve tornado clock precision
Why:
The single-precision tornado clock stopped advancing at the first 1/120-second float stall boundary, freezing lifecycle and presentation state.

Ownership:
Gameplay TornadoSystem is the lifecycle clock owner; Runtime Replay and Prediction own their serialized boundaries, and TornadoVisualPass owns visual phase reduction.

What:
Retain ordinary float arithmetic through the compatibility range, continue in double beyond the stall boundary, propagate the precise clock through solver snapshot v5 and prediction raw-v5/sectioned-v6, preserve historical float layouts, and add focused exact-restore/archive/visual evidence.

Validation:
deferred by explicit owner direction; no build, test, scanner, or validation command ran.

Baselines/Artifacts:
No baselines or generated artifacts changed.

Review:
Four serial read-only rubber-duck passes completed; the final verdict is CLEAN. Historical sectioned-v4 prediction and solver-v2-v4 byte fixtures remain residual static coverage risk.

+++ SkullbonezSource/Gameplay/TornadoGameplay.cpp
@@ -166,7 +166,7 @@ const TornadoSystemConfig& TornadoGameplay::GetSystemConfig() const
     return m_system.GetConfig();
 }
 
-float TornadoGameplay::GetSystemElapsedSeconds() const
+double TornadoGameplay::GetSystemElapsedSeconds() const
 {
     return m_system.GetElapsedSeconds();
 }
@@ -227,7 +227,7 @@ void TornadoGameplay::SetFieldValue( float TornadoFieldConfig::* field, float va
 
 void TornadoGameplay::SetReplayState( const std::vector<float>& captureSeconds,
                                       const std::vector<float>& ejectCooldownSeconds, const TornadoFieldConfig& fieldConfig,
-                                      const TornadoSystemConfig& systemConfig, float systemElapsedSeconds )
+                                      const TornadoSystemConfig& systemConfig, double systemElapsedSeconds )
 {
     if ( captureSeconds.size() > m_captureSeconds.capacity() ||
          ejectCooldownSeconds.size() > m_ejectCooldownSeconds.capacity() )
